🛠[프로젝트_새 프로젝트] : 새 프로젝트 컨테이너 생성
